pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "datajournals"
version = "0.8.4"
authors = [
    {name="Kozmik Moore", email="koz.moore@gmail.com"},
]
description = "A web-based journaling application."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Framework :: Flask",
]
dependencies = [
    "Flask==2.3.2",
    'Flask-Login==0.6.2',
    'email-validator==2.0.0.post2',
    "Flask-Markdown==0.3",
    "Flask-Migrate==4.0.4",
    "Flask-SQLAlchemy==3.0.3",
    "Flask-WTF==1.1.1",
    "gunicorn==20.1.0",
    "numpy~=1.25.2",
    "pandas~=2.0.3",
    "seaborn~=0.13.0",
]

[tool.setuptools.packages.find]
exclude = ["instance"]